🗃Refactor: component export 방식 변경 및 네이밍 수정
