+++ pages/_app.js
@@ -1,7 +1,14 @@
-import '../styles/globals.css'
+import '../styles/globals.css';
+import { RecoilRoot } from 'recoil';
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return (
+    <>
+      <RecoilRoot>
+        <Component {...pageProps} />
+      </RecoilRoot>
+    </>
+  );
 }
 
-export default MyApp
+export default MyApp;
