+++ src/components/index.js
@@ -1,0 +1,22 @@
+import { Flex } from './Flex/Flex';
+
+// Modal related
+import { Modal } from './Modal/Modal';
+import { ModalFooter } from './Modal/ModalFooter/ModalFooter';
+import { ModalMain } from './Modal/ModalMain/ModalMain';
+import { ModalProfile } from './Modal/ModalProfile/ModalProfile';
+import { ModalHeader } from './Modal/ModalHeader/ModalHeader';
+
+// Home related
+import { Hero } from './Home/Hero';
+import { Animation } from './Home/Animation';
+
+// Commons
+import Footer from './footer';
+import Layout from './layout';
+import Header from './header';
+
+// DarkMode
+import DarkModeToggleButton from './DarkMode';
+
+export { Flex, Modal, ModalFooter, ModalMain, ModalProfile, ModalHeader, Hero, Animation, Footer, Layout, Header, DarkModeToggleButton };
