+++ src/states/index.js
@@ -1,0 +1,3 @@
+import modalState from './Modal';
+
+export { modalState };
